extracted mergeStores helper
