+++ core/lib/module/helpers.ts
@@ -1,4 +1,8 @@
 import { VueStorefrontModuleConfig } from './types'
+import { Module } from 'vuex'
+import merge from 'lodash-es/merge'
+import some from 'lodash-es/some'
+import find from 'lodash-es/find'
 
 function doesStoreAlreadyExists (key: string, registeredModules: VueStorefrontModuleConfig[]) : boolean {
   registeredModules.forEach(m => {
@@ -9,6 +13,24 @@ function doesStoreAlreadyExists (key: string, registeredModules: VueStorefrontMo
   return false
 }
 
+
+function mergeStores (
+  originalStore: { modules?: { key: string, module: Module<any, any> }[], plugin?: Function }, 
+  extendedStore:  { modules?: { key: string, module: Module<any, any> }[], plugin?: Function }
+) {
+  let mergedArray = []
+  originalStore.modules.map(item => {
+    mergedArray.push(merge(item, find(extendedStore.modules, { 'key' : item.key })));
+  })
+  extendedStore.modules.map(extendedStoreItem => {
+    if(some(originalStore.modules, null, { 'key' : extendedStoreItem.key}) === false){
+      mergedArray.push(extendedStoreItem)
+    }
+  })
+  return mergedArray
+}
+
 export {
-  doesStoreAlreadyExists
+  doesStoreAlreadyExists,
+  mergeStores
 }
